requirements.txt
google-generativeai
pyautogui
mss
pillow
rich
